# swagger-userManagement.json
{"openapi":"3.0.3","info":{"title":"Api Documentation","description":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"},"version":"1.0"},"servers":[{"url":"http://localhost:80","description":"Inferred Url"}],"tags":[{"name":"basic-error-controller","description":"Basic Error Controller"},{"name":"create-role-controller","description":"Create Role Controller"},{"name":"create-user-controller","description":"Create User Controller"},{"name":"delete-role-controller","description":"Delete Role Controller"},{"name":"delete-user-controller","description":"Delete User Controller"},{"name":"operation-handler","description":"Operation Handler"},{"name":"read-role-controller","description":"Read Role Controller"},{"name":"update-role-controller","description":"Update Role Controller"},{"name":"user-controller","description":"User Controller"},{"name":"web-mvc-links-handler","description":"Web Mvc Links Handler"}],"paths":{"/error":{"get":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingGET","responses":{"200":{"description":"OK","content":{"text/html":{"schema":{"$ref":"#/components/schemas/ModelAndView"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPUT","responses":{"200":{"description":"OK","content":{"text/html":{"schema":{"$ref":"#/components/schemas/ModelAndView"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPOST","responses":{"200":{"description":"OK","content":{"text/html":{"schema":{"$ref":"#/components/schemas/ModelAndView"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingDELETE","responses":{"200":{"description":"OK","content":{"text/html":{"schema":{"$ref":"#/components/schemas/ModelAndView"}}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"options":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingOPTIONS","responses":{"200":{"description":"OK","content":{"text/html":{"schema":{"$ref":"#/components/schemas/ModelAndView"}}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"head":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingHEAD","responses":{"200":{"description":"OK","content":{"text/html":{"schema":{"$ref":"#/components/schemas/ModelAndView"}}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"patch":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPATCH","responses":{"200":{"description":"OK","content":{"text/html":{"schema":{"$ref":"#/components/schemas/ModelAndView"}}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"trace":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingTRACE","responses":{"200":{"description":"OK","content":{"text/html":{"schema":{"$ref":"#/components/schemas/ModelAndView"}}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/api/role":{"post":{"tags":["create-role-controller"],"summary":"create","operationId":"createUsingPOST","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleControllerIn"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateRoleControllerOut"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/user":{"post":{"tags":["create-user-controller"],"summary":"create","operationId":"createUsingPOST_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserControllerIn"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateUserControllerOut"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/role/{uuid}":{"put":{"tags":["update-role-controller"],"summary":"update","operationId":"updateUsingPUT","parameters":[{"name":"uuid","in":"path","description":"uuid","required":true,"style":"simple","schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleControllerIn"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UpdateRoleControllerOut"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["delete-role-controller"],"summary":"delete","operationId":"deleteUsingDELETE","parameters":[{"name":"uuid","in":"path","description":"uuid","required":true,"style":"simple","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/api/user/{uuid}":{"delete":{"tags":["delete-user-controller"],"summary":"delete","operationId":"deleteUsingDELETE_1","parameters":[{"name":"uuid","in":"path","description":"uuid","required":true,"style":"simple","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/actuator/health":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/health/**":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_1","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/info":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_2","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/roles/find":{"get":{"tags":["read-role-controller"],"summary":"roles","operationId":"rolesUsingGET","requestBody":{"content":{"application/json":{"schema":{"type":"string"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/backend/":{"get":{"tags":["user-controller"],"summary":"home","operationId":"homeUsingGET","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator":{"get":{"tags":["web-mvc-links-handler"],"summary":"links","operationId":"linksUsingGET","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}},"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"components":{"schemas":{"CreateRoleControllerIn":{"title":"CreateRoleControllerIn","type":"object","properties":{"priority":{"type":"integer","format":"int32"},"roleName":{"type":"string"}}},"CreateRoleControllerOut":{"title":"CreateRoleControllerOut","type":"object","properties":{"priority":{"type":"integer","format":"int32"},"roleName":{"type":"string"},"uuid":{"type":"string"}}},"CreateUserControllerIn":{"title":"CreateUserControllerIn","type":"object","properties":{"createUserIn":{"$ref":"#/components/schemas/CreateUserIn"}}},"CreateUserControllerOut":{"title":"CreateUserControllerOut","type":"object","properties":{"uuid":{"type":"string"}}},"CreateUserIn":{"title":"CreateUserIn","type":"object","properties":{"email":{"$ref":"#/components/schemas/Email"},"password":{"$ref":"#/components/schemas/Password"},"userName":{"$ref":"#/components/schemas/UserName"},"uuid":{"$ref":"#/components/schemas/Uuid"}}},"Email":{"title":"Email","type":"object","properties":{"email":{"type":"string"}}},"Link":{"title":"Link","type":"object","properties":{"href":{"type":"string"},"templated":{"type":"boolean"}}},"ModelAndView":{"title":"ModelAndView","type":"object","properties":{"empty":{"type":"boolean"},"model":{"type":"object"},"modelMap":{"type":"object","additionalProperties":{"type":"object"}},"reference":{"type":"boolean"},"status":{"type":"string","enum":["ACCEPTED","ALREADY_REPORTED","BAD_GATEWAY","BAD_REQUEST","BANDWIDTH_LIMIT_EXCEEDED","CHECKPOINT","CONFLICT","CONTINUE","CREATED","DESTINATION_LOCKED","EXPECTATION_FAILED","FAILED_DEPENDENCY","FORBIDDEN","FOUND","GATEWAY_TIMEOUT","GONE","HTTP_VERSION_NOT_SUPPORTED","IM_USED","INSUFFICIENT_SPACE_ON_RESOURCE","INSUFFICIENT_STORAGE","INTERNAL_SERVER_ERROR","I_AM_A_TEAPOT","LENGTH_REQUIRED","LOCKED","LOOP_DETECTED","METHOD_FAILURE","METHOD_NOT_ALLOWED","MOVED_PERMANENTLY","MOVED_TEMPORARILY","MULTIPLE_CHOICES","MULTI_STATUS","NETWORK_AUTHENTICATION_REQUIRED","NON_AUTHORITATIVE_INFORMATION","NOT_ACCEPTABLE","NOT_EXTENDED","NOT_FOUND","NOT_IMPLEMENTED","NOT_MODIFIED","NO_CONTENT","OK","PARTIAL_CONTENT","PAYLOAD_TOO_LARGE","PAYMENT_REQUIRED","PERMANENT_REDIRECT","PRECONDITION_FAILED","PRECONDITION_REQUIRED","PROCESSING","PROXY_AUTHENTICATION_REQUIRED","REQUESTED_RANGE_NOT_SATISFIABLE","REQUEST_ENTITY_TOO_LARGE","REQUEST_HEADER_FIELDS_TOO_LARGE","REQUEST_TIMEOUT","REQUEST_URI_TOO_LONG","RESET_CONTENT","SEE_OTHER","SERVICE_UNAVAILABLE","SWITCHING_PROTOCOLS","TEMPORARY_REDIRECT","TOO_EARLY","TOO_MANY_REQUESTS","UNAUTHORIZED","UNAVAILABLE_FOR_LEGAL_REASONS","UNPROCESSABLE_ENTITY","UNSUPPORTED_MEDIA_TYPE","UPGRADE_REQUIRED","URI_TOO_LONG","USE_PROXY","VARIANT_ALSO_NEGOTIATES"]},"view":{"$ref":"#/components/schemas/View"},"viewName":{"type":"string"}}},"Password":{"title":"Password","type":"object","properties":{"password":{"type":"string"}}},"UpdateRoleControllerIn":{"title":"UpdateRoleControllerIn","type":"object","properties":{"priority":{"type":"integer","format":"int32"},"roleName":{"type":"string"}}},"UpdateRoleControllerOut":{"title":"UpdateRoleControllerOut","type":"object","properties":{"priority":{"type":"integer","format":"int32"},"roleName":{"type":"string"},"uuid":{"type":"string"}}},"UserName":{"title":"UserName","type":"object","properties":{"userName":{"type":"string"}}},"Uuid":{"title":"Uuid","type":"object","properties":{"uuid":{"type":"string"}}},"View":{"title":"View","type":"object","properties":{"contentType":{"type":"string"}}}}}}